test: reservationService 테스트 추가
